+++ lib/src/languages/java/java11/java11_grammar.dart
@@ -475,6 +475,7 @@ class Java11GrammarDefinition extends Java11GrammarLexer {
   Parser<ASTExpression> expressionNoOperation() =>
       (expressionNegate() |
               expressionLiteral() |
+              expressionGroupFunctionInvocation() |
               expressionGroup() |
               expressionListLiteral() |
               expressionListEmptyLiteral() |
@@ -485,7 +486,8 @@ class Java11GrammarDefinition extends Java11GrammarLexer {
               expressionFunctionInvocation() |
               expressionVariableEntryAccess() |
               expressionNullValue() |
-              expressionVariableAccess())
+              expressionVariableAccess() |
+              expressionNegative())
           .cast<ASTExpression>();
 
   Parser<ASTExpressionNegation> expressionNegate() =>
@@ -496,10 +498,34 @@ class Java11GrammarDefinition extends Java11GrammarLexer {
             return ASTExpressionNegation(exp);
           });
 
+  Parser<ASTExpressionNegative> expressionNegative() =>
+      (char('-').trimHidden() &
+              (ref0(expressionNoOperation) | ref0(expressionGroup)))
+          .map((v) {
+            var exp = v[1] as ASTExpression;
+            return ASTExpressionNegative(exp);
+          });
+
   Parser<ASTExpression> expressionGroup() =>
       (char('(').trimHidden() & ref0(expression) & char(')').trimHidden()).map(
         (v) => v[1] as ASTExpression,
       );
+
+  Parser<ASTExpressionGroupFunctionInvocation>
+  expressionGroupFunctionInvocation() =>
+      (ref0(expressionGroup) &
+              char('.') &
+              identifier() &
+              char('(').trimHidden() &
+              ref0(expressionSequence).optional() &
+              char(')').trimHidden())
+          .map((v) {
+            var expression = v[0] as ASTExpression;
+            var name = v[2] as String;
+            var args = v[4] as List<ASTExpression>?;
+            args ??= <ASTExpression>[];
+            return ASTExpressionGroupFunctionInvocation(expression, name, args);
+          });
 
   Parser<ASTExpressionFunctionInvocation> expressionFunctionInvocation() =>
       ((identifier() & char('.')).optional() &